svd!: allow lower bidiagonal
Close JuliaLinearAlgebra/TSVD.jl#33

TSVD requires the SVD of a lower bidiagonal matrix.
If the element type is not one supported by LAPACK and if
GenericLinearAlgebra is imported, the call dispatches to it.
However, GenericLinearAlgebra only implements the SVD of an *upper*
bidiagonal.
This PR introduces a helper function to pass the transposed of the
bidiagonal if the latter is a lower bidiagonal.
As a result, the truncated SVD of a matrix of, say, BigFloat, is now
possible.